pyproject.toml
[tool.poetry]
name = "gprofiler_custom_gmt"
version = "0.1.0"
description = ""
authors = ["Hugo Martiniano <hugomartiniano@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
requests = "^2.25.1"
gprofiler-official = "^1.0.0"
pandas = "^1.2.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gprofiler_custom_gmt = "gprofiler_custom_gmt.main:cli"

